Better control
PI controller on odometry, regulated in speed. Distance is integrated
(and multiplied with the KI term).
Must add some kind of odometer reset, since now the PIDPOD is stable
only in a particular point on a straight line.
Maybe add odometry_total*KI to the acc_value when the PIDPOD is fairly
stable and then reset odometry_total to 0?

The wifi is still buggy! Do not use wifi

+++ Software/Energia/libraries/angle_control_loop/imu_control.cpp
@@ -118,6 +118,8 @@ void ControllerIntHandler(void)
   	/* Compute controller speed */
   	speed = angle * Kp + integral * Ki + gyro_angle * Kd;
   	
+  	
+  	
   	/* Apply command value to the motors (as long as the DIP4 is = 1) */
   	if(digitalRead(DIP4) && angle_acceptable())
     	setSpeed(speed, speed);
@@ -133,7 +135,7 @@ float get_accelerometer_default(void)
 	return accelerometer;
 }
 
-int8_t get_speed(void)
+int16_t get_speed(void)
 {
 	return  speed;
 }
