+++ Software/Energia/libraries/angle_control_loop/imu_control.h
@@ -18,7 +18,7 @@ void ControllerIntHandler(void);
 float get_accelerometer_default(void);
 
 /* Return speed value */
-int8_t get_speed(void);
+int16_t get_speed(void);
 
 /* Set controller Parameters */
 void set_controller_parameters(float p, float i, float d);
